Use CWSDSTUB.EXE to produce standalone executables
CWSDSTUB.EXE includes a copy of CWSDPMI and loads the DJGPP COFF that is
appended to it. The DJGPP toolchain is still not needed; COFFs are
produced by a linker script following the spec:
http://www.delorie.com/djgpp/doc/coff/.

Since executables are standalone, DOSBox (the most popular DOS emulator)
can now be used to run the demo program.

I'm quite proud of the work I've done on the ELF loader, but it is no
longer needed.