Align and pad program sections to fix heisenbugs
Printed strings sometimes had return and newline characters cut off,
resulting in corrupted output. Altering the contents of the program
would make the problem come and go. I created a custom build of DOSBox
to observe the contents of the transfer buffer when the read and write
DOS API (int 21h) was used. The trailing characters were not always
present, yet they could be seen in the executable. The affected string
was at the end of the executable, and I realized that the trailing bytes
were being dropped at program load time if the section was not aligned
to 4 bytes. Aligning and padding the program sections to 4 bytes seems
to resolve the issue. I could not find an authoritative document that
specifies the COFF alignment requirements. I only found COFF headers
with comments on struct fields mentioning that the section size should
be padded to the forward boundary.